Restore part of the project and WIP RL agent train
Many modules that were unused would also be unusable, as the import structure was destroyed by the previous project refactor. Some of it was restored, in particular, enough of it to make nlns.main.run_train usable for training of the RL agent. It is now possible to run it through:
python -m nlns.main.run_train [options...]

It can only train the RL agent, badly (probably there are still some compatibility issues since I patched it brainlessly). In the near future it will be imperative to find a proper way to train both main neural operators together (OR drop this idea entirely).

All main scripts shall finally be put inside the main package and designed to be invoked through their namespace (python -m).