Add a special function for extra requirements
Some modules contain features that rely on third party software. However, the library itself does not need these extra modules to work. A special module_found(...) function introduced, that can be used to check if a module is importable and raise a clean exception that instructs the user about a missing extra dependency.