More flexible requirements
Installing the library shall now be easier. PySCIPOpt must be explicitly installed in order to use the related operators.

NB: torch-scatter (and possibly other torch expansions) require torch to be installed a priori. A clean install of the enviroment would then be:

pip install torch
pip install -r requirements.txt

Installing torch and its extensions within the same call is not possible as they try to import torch directly, and the package index is not updated until all the packages are collected and built.