Add RNG retrieval function
In the future it will probably be able to support numpy states as well. Retrieves a simple random.Random object properly seeded. None values will return a package global rng (nlns.default_rng), so that every random operation in the library can make use of it, preventing the dangerous usage of a global state.