RLAgentRepair now accepts plain VRPSolutions
Instead of VRPNeuralSolutions. Neural solutions are built at the moment. Possibly in the future it could be a good idea to make them encapsulate solutions instead of being a subclass. The subclass is "necessary" to update the neural routes when a standard heuristic operator is used. Since we now build them on demand, this feature is not longer needed and we can simplify the neural solution object accordingly. Performance is still comparable to the old setup, as the entire representation was rebuilt at each step anyway.

Standard unit tests are still missing due to some technicalities.