Use numpy for uchoa generation (not torch)
Random functions for numpy feel more natural (instead of doing a [0, 1) random and multiplying it to get uniforms and such). Moreover, this is one of the final steps in decoupling torch from the entire project, so that a minimal installation will not require pytorch.

This also include reproducibility (tested and working).

Also fixed a bug in said generation, where a vector of capacities would be generated (one per instance) but not correctly unpacked (each instance would get the entire vector).

Performance are comparable to the previous implementation.