Change LNS history mechanism to output a dict
Similarly to other frameworks histories (e.g. keras) the history is a
dict containing a sequence of logged statistics per key. This will make
it easier to change/add/generalize them in the future if we want.

Current time and current best cost are currently shown. "Current best
cost" is the cost of the currently "accepted" solution, which based on
the acceptance criterion may or may not be the overall best solution
found so far.